requirements.txt
streamlit>=1.28.0
requests>=2.31.0
pandas>=2.0.0
python-dotenv>=1.0.0
plotly>=5.17.0
matplotlib>=3.7.0
pytest>=7.4.0